Unified Neighbourhood + clean Random Local Solver
`available_moves` should now be passed as a parameter when creating Neighbourhood.

Funtionality of NA and NB was moved to regular Neighbourhood. This allows Random Local Solver to use only one Neighbourhood instead of two.

Random Local Solver was also slightly changed in terms of picking a move to make.

NA and NB classes were deemed to be redundant and thus removed.